fix(security): wire redact + SSRF default-deny + BM25 overflow + daemon start cleanup
Second audit uncovered that the prior 'redact' package was entirely unused —
every handler was writing raw tool arguments and output straight to the
journal. Plus four other critical issues.

C1 transport: wire the redactor through NewHandlers. Remember/Exec/Read/Fetch
    now redact Data maps, Tags, Stdout/Stderr/Body/Summary, and
    ContentMatch.Text before the journal or the MCP client sees them.
    Regression test TestHandlersRedactRemember confirms tokens written via
    Remember do not survive in journal.jsonl.

C2 sandbox: assertFetchURLAllowed used to return nil on DNS failure, allowing
    an attacker-controlled domain that briefly NXDOMAIN'd to slip through the
    SSRF guard when the HTTP client's resolver later succeeded. DNS failure
    is now a hard block. Bare IP literals bypass DNS entirely (still subject
    to isBlockedIP).

C3 core(index): PostingList.TFs was uint16; a token repeated 65 536+ times
    silently wrapped and corrupted BM25 scores. Widened to uint32 through
    Add and Remove. Memory overhead is trivial.

C4 daemon(start): Start now uses CompareAndSwap to flip the running flag so
    two concurrent Start calls cannot both bind the listener. When any start
    phase returns an error or panics, any already-opened resource
    (server listener, fswatcher goroutine) is rolled back before Start
    returns — previously a panic after server.Start left the TCP port /
    Unix socket dangling.

C5 config: Load() now runs Validate() on the merged result. Typo'd
    durability/idle_timeout values fail at daemon start instead of silently
    degrading to defaults.

Tests: go test ./... passes; go vet clean.

+++ internal/core/index.go
@@ -9,9 +9,12 @@ import (
 )
 
 // PostingList holds the document IDs and term frequencies for a term.
+// TFs is uint32: uint16 would overflow on tokens repeated 65 536+ times
+// (e.g. a huge log containing the same identifier over and over), silently
+// corrupting BM25 scores.
 type PostingList struct {
 	IDs []string // ULIDs, sorted
-	TFs []uint16 // term frequencies, parallel to IDs
+	TFs []uint32 // term frequencies, parallel to IDs
 }
 
 // Index implements an in-memory inverted index with BM25 scoring.
@@ -105,7 +108,7 @@ func (ix *Index) Add(e Event) {
 			ix.stemPL[stem] = pl
 		}
 		pl.IDs = append(pl.IDs, id)
-		pl.TFs = append(pl.TFs, uint16(freq))
+		pl.TFs = append(pl.TFs, uint32(freq))
 	}
 
 	// Also index for trigram search
@@ -246,7 +249,7 @@ func (ix *Index) Remove(id string) {
 	// Remove from stem posting lists.
 	for stem, pl := range ix.stemPL {
 		newIDs := make([]string, 0, len(pl.IDs))
-		newTFs := make([]uint16, 0, len(pl.TFs))
+		newTFs := make([]uint32, 0, len(pl.TFs))
 		for i, docID := range pl.IDs {
 			if docID != id {
 				newIDs = append(newIDs, docID)
